# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Todo API
  version: 1.0.0
  description: API for managing smart todos with OIDC authentication
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.smarttodo.com
    description: Production server
paths:
...
  /api/v1/todos:
    get:
      summary: List todos
      description: Get all todos for the authenticated user, optionally filtered by time_horizon and status
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: time_horizon
          in: query
          description: Filter by time horizon
          schema:
            type: string
            enum: [next, soon, later]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, processing, completed]
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodosResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new todo
      description: Create a new todo item
      tags:
        - Todos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/todos/{id}:
    get:
      summary: Get a todo by ID
      description: Retrieve a specific todo by its ID
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Update a todo
      description: Update an existing todo
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete a todo
      description: Delete a todo by ID
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Todo deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/todos/{id}/complete:
    post:
      summary: Mark todo as completed
      description: Mark a todo as completed
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Todo ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Todo marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/todos/tags/stats:
    get:
      summary: Get tag statistics
      description: Returns aggregated tag statistics for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tag statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from OIDC provider
  schemas:
...
    CreateTodoRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
    UpdateTodoRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
        time_horizon:
          type: string
          enum: [now, soon, later]
        status:
          type: string
          enum: [pending, processing, completed]
    Todo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        text:
          type: string
        time_horizon:
          type: string
          enum: [now, soon, later]
        status:
          type: string
          enum: [pending, processing, completed]
        metadata:
          $ref: '#/components/schemas/Metadata'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    Metadata:
      type: object
      properties:
        category_tags:
          type: array
          items:
            type: string
        priority:
          type: string
        context:
          type: array
          items:
            type: string
        duration:
          type: string
    TodoResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Todo'
        timestamp:
          type: string
          format: date-time
    TodosResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
        timestamp:
          type: string
          format: date-time
    TagStatsResponse:
      type: object
      properties:
        tag_stats:
          type: object
          additionalProperties:
            type: object
            properties:
              total:
                type: integer
                description: Total count of todos with this tag
              ai:
                type: integer
                description: Count of todos where tag was AI-generated
              user:
                type: integer
                description: Count of todos where tag was user-defined
        tainted:
          type: boolean
          description: Whether the statistics are being recomputed
        last_analyzed_at:
          type: string
          format: date-time
          nullable: true
          description: When the statistics were last analyzed
...
    Error:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'